requirements.txt
# CBAM-STN-TPS-YOLO Requirements
# Core Deep Learning Framework
torch>=1.12.0,<2.1.0
torchvision>=0.13.0,<0.16.0
torchaudio>=0.12.0,<2.1.0

# Computer Vision & Image Processing
opencv-python>=4.6.0,<5.0.0
Pillow>=9.0.0,<10.0.0
albumentations>=1.3.0,<2.0.0
scikit-image>=0.19.0,<1.0.0

# Scientific Computing
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<3.0.0
matplotlib>=3.6.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Required for TPS mathematical operations
scipy>=1.9.0,<2.0.0

# Machine Learning
scikit-learn>=1.1.0,<2.0.0
ultralytics>=8.0.0,<9.0.0

# Data Processing & Configuration
PyYAML>=6.0,<7.0
tqdm>=4.64.0,<5.0.0
h5py>=3.7.0,<4.0.0
click>=8.0.0,<9.0.0

# Visualization & Monitoring
plotly>=5.10.0,<6.0.0
wandb>=0.13.0,<1.0.0
tensorboard>=2.10.0,<3.0.0

# Model Export & Deployment
onnx>=1.12.0,<2.0.0
onnxruntime>=1.12.0,<2.0.0

# Jupyter Notebook Support
jupyter>=1.0.0,<2.0.0
ipywidgets>=8.0.0,<9.0.0
ipykernel>=6.15.0,<7.0.0

# Development Tools (Optional)
pytest>=7.0.0,<8.0.0
black>=22.0.0,<23.0.0
flake8>=5.0.0,<6.0.0
isort>=5.10.0,<6.0.0
mypy>=0.991,<1.0

# Additional Utilities
rich>=12.0.0,<13.0.0
typer>=0.6.0,<1.0.0

# GPU Acceleration (Install based on CUDA version)
# For CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# For CPU only: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

# Optional Dependencies (install separately if needed)
# tensorrt>=8.4.0  # For NVIDIA deployment
# openvino>=2022.2.0  # For Intel deployment
# imgaug>=0.4.0  # Additional augmentations
# kornia>=0.6.0  # Differentiable computer vision